fix: split oversized files into components, link brand page from settings
Extract landing page sections into components/landing/ (729->156 lines).
Extract specialist profile sub-components into components/specialist/ (521->257 lines).
Add HeaderBack to brand.tsx + dev-only "Design System" link in settings.
