+++ components/landing/mockData.ts
@@ -1,0 +1,52 @@
+export const MOCK_SPECIALISTS = [
+  { name: "Алексей М.", rating: 4.9, tag: "Выездная проверка" },
+  { name: "Ирина К.", rating: 4.8, tag: "Камеральная проверка" },
+  { name: "Дмитрий С.", rating: 5.0, tag: "ОКК" },
+];
+
+export const PROBLEMS = [
+  {
+    icon: "\u26A0\uFE0F",
+    title: "Не знаете, насколько серьёзно",
+    text: "Штраф? Доначисления? Блокировка счёта? Без опыта сложно оценить масштаб проблемы и последствия.",
+  },
+  {
+    icon: "\uD83D\uDCCB",
+    title: "Не знаете что делать",
+    text: "Куда идти, какие документы готовить, как отвечать на требования инспекции — непонятно с чего начать.",
+  },
+  {
+    icon: "\uD83D\uDD0D",
+    title: "Не можете найти нужного",
+    text: "Обычные бухгалтеры не разбираются в P2P. А те, кто разбирается — их ещё нужно найти в вашем городе.",
+  },
+];
+
+export const FEATURES = [
+  {
+    icon: "\uD83D\uDDFA\uFE0F",
+    title: "Фильтр по городу и ФНС",
+    text: "Только специалисты, которые работают с вашей инспекцией.",
+  },
+  {
+    icon: "\u2709\uFE0F",
+    title: "Специалисты пишут первыми",
+    text: "Не нужно обзванивать всех. Они изучают и приходят к вам.",
+  },
+  {
+    icon: "\uD83D\uDCB0",
+    title: "Бесплатно для клиента",
+    text: "Создать заявку и получить предложения ничего не стоит.",
+  },
+  {
+    icon: "\uD83D\uDD00",
+    title: "Два пути входа",
+    text: "Создать заявку или найти специалиста в каталоге и написать напрямую.",
+  },
+];
+
+export const STEPS = [
+  { num: "1", label: "Создайте заявку" },
+  { num: "2", label: "Специалисты напишут сами" },
+  { num: "3", label: "Выберите и решите вопрос" },
+];
