+++ components/specialist/types.ts
@@ -1,0 +1,43 @@
+export interface FnsServiceGroup {
+  fns: { id: string; name: string; code: string };
+  city: { id: string; name: string };
+  services: { id: string; name: string }[];
+}
+
+export interface SpecialistProfile {
+  description: string | null;
+  phone: string | null;
+  telegram: string | null;
+  whatsapp: string | null;
+  officeAddress: string | null;
+  workingHours: string | null;
+}
+
+export interface ContactMethodItem {
+  id: string;
+  type: string;
+  value: string;
+  label: string | null;
+  order: number;
+}
+
+export interface SpecialistDetail {
+  id: string;
+  firstName: string | null;
+  lastName: string | null;
+  avatarUrl: string | null;
+  isAvailable: boolean;
+  createdAt: string;
+  profile: SpecialistProfile | null;
+  fnsServices: FnsServiceGroup[];
+  requestsCount?: number;
+}
+
+export interface SimilarSpecialist {
+  id: string;
+  firstName: string | null;
+  lastName: string | null;
+  avatarUrl: string | null;
+  services: { id: string; name: string }[];
+  cities: { id: string; name: string }[];
+}
